requirements.txt

pywebview
requests
